Refactor core routing logic to utilize new authService
This change focuses on two things: Using the new authentication system and
ensuring that static redirection still occurs in production mode code. To
achieve this, a full rewrite of the router interceptor lifecycles had to
be done.

Note that this implementation allows any route to handle the OIDC callback
flow. This may or may not be best practices, as other examples frequently
direct the OIDC provider to redirect back to a known hard path instead. We
went this generic approach as we want to reasonably preserve the potential
differences in path and redirections that will also already happen in
tandem with on-load behavior.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>