Refactor main initializeApp to force a pinia configStore init
While the service singletons are up and running at this point, we want the
stores to also be aware of the data changes and be able to react to async
update events that show up in userManager. For this to happen, the init
function must explicitly add event handlers to userManager before the app
starts up.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>