Update authService to use a static userManager instance
This change reduces instance duplication and makes sure the event handlers
mounted in the store stay mounted to the same userManager. This change also
adds debug support to conditionally expose the direct state elements in the
setup function. Lastly, InitApp is removed as we now initialize earlier in
the lifecycle before vue even gets mounted to the DOM.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>