Attempt to perform vue-router navigation after /oidc/callback
Unfortunately this doesn't appear to fully navigate the application. Left
some commented breadcrumbs to revisit later. Also added some various fixes
to ensure the application is nominally working on quick inspection. Will
need to revisit how and where userStore actually needs to be loaded, but
leaving it in top level of application to maintain functionality.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>