Restructure vue-router and views to leverage nested router-view mounts
This change ensures that we are properly leveraging nested router
definitions correctly and that it shows up on the vue tools inspector as
intended. It also does a first crack at ensuring that only route points
that require query parameter handling leverage the createProps helper.
Lastly, a duplicate route definition is removed.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>