Extend authStore to acquire userId on _updateState trigger
In general, an authenticated user will need to know the database userId in
order to perform subsequent business logic. Instead of depending on a
separate userstore to cache this information, we instead let the
authentication system handle this for us as the userId's lifecycle is the
same as the logged-in user's session. We also tack on a few more copy
buttons to the developer view.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>