Make conditionals for toggle more readable and swap their order
The conditional for disabling the public toggle has been rearranged - from NOT x OR NOT y, to NOT (x AND y) - to make it a bit more readable and slightly less confusing due to the use of multiple double negatives.

The order of those conditionals has also been swapped, so that we can potentially avoid an additional network call.