ObjectTable: select all on current page only
The default behaviour of the DataTable "select all" checkbox will select all items in the table, including those not on the current page. This has been overridden.

Due to the limitations of DataTable emits (they don't consistently return a list of objects sorted in the same order as they're displayed), document.querySelector() is used to directly find out what's actually on the current DataTable page.