Force config call to never cache
There's a possibility that certain browser environments are encountering
cached 304 responses when asking for the config. While this is correct
serverside, on the client if it is on a new tab, the value will not be
cached, and thus may end up being undefined. This header change hopefully
makes it so that it will always return a 200 with body instead of a network
optimized 304.

Signed-off-by: Jeremy Ho <jujaga@gmail.com>