Update bucket/object permissions: limit scope of post-update fetch call
When adding or removing permissions for a bucket/object, a call is made to fetch permissions immediately after. This call was not scoped at all, so it would fetch permissions for ALL buckets/objects, resulting in a very heavy COMS API call. This has now been restricted to just the bucket/object being updated.